Updated and organized hardware
New versions of the sumo components have been included. New wheel prevents the shaft from slipping, the base has new holes for attaching the PCB and the SHARP sensors in a beter position, and the engine bracket includes a gap to keep the encoder cables tight to the encoder.

New organization of the differentt version, including printables has been done.